Added ObjectName to OutboundEvent
In certain cases the ManagedObjectReference alone might not be useful,
e.g. when an object is deleted.

This commit adds a new field `objectName` to the outbound JSON data
which serves as additional metadata for the receiver.

Note: ObjectName for certain vSphere objects, e.g. virtual machines, is
only unique within a folder (if applicable) as per the description
[here](https://docs.vmware.com/en/VMware-vSphere/6.7/com.vmware.vsphere.vm_admin.doc/GUID-76E73C62-A973-4839-BB67-AC1817908E6D.html).

This added field should be treated as metadata and not for requests
against vCenter (use `ManagedObjectReference` instead).

Signed-off-by: Michael Gasch <embano1@live.com>